requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
beanie>=1.23.0
motor>=3.3.0
pymongo>=4.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# PDF Generation
weasyprint>=60.0

# CORS is built into FastAPI, no separate package needed

# Environment variables
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Date/Time utilities
python-dateutil>=2.8.2
